getTempTablePrefix {SqlRender}	R Documentation

Get the prefix used for emulated temp tables for DBMSs that do not support temp tables (e.g. Oracle, BigQuery).

Description

Get the prefix used for emulated temp tables for DBMSs that do not support temp tables (e.g. Oracle, BigQuery).

Usage

getTempTablePrefix()

Value

The prefix string.

Examples

getTempTablePrefix()

[Package SqlRender version 1.13.1 Index]