launchSqlRenderDeveloper {SqlRender}	R Documentation

Launch the SqlRender Developer Shiny app

Description

Launch the SqlRender Developer Shiny app

Usage

launchSqlRenderDeveloper(launch.browser = TRUE)

Arguments

launch.browser
Should the app be launched in your default browser, or in a Shiny window. Note: copying to clipboard will not work in a Shiny window.

Details

Launches a Shiny app that allows the user to develop SQL and see how it translates to the supported dialects.

[Package SqlRender version 1.13.1 Index]