renderSql

renderSql Renders SQL code based on parameterized SQL and parameter values.

renderSql(sql = "", ...)

Arguments

sql
The parameterized SQL

...
Parameter values

Value

A list containing the following elements:

parameterizedSql

The original parameterized SQL code

sql

The rendered sql

Details

This function takes parameterized SQL and a list of parameter values and renders the SQL that can be send to the server. Parameterization syntax:

@parameterName

Parameters are indicated using a @ prefix, and are replaced with the actual values provided in the renderSql call.

{DEFAULT @parameterName = parameterValue}

Default values for parameters can be defined using curly and the DEFAULT keyword.

{if}?{then}:{else}

The if-then-else pattern is used to turn on or off blocks of SQL code.

Examples

renderSql("SELECT * FROM @a;", a = "myTable")

#> $originalSql
#> [1] "SELECT * FROM @a;"
#> 
#> $sql
#> [1] "SELECT * FROM myTable;"
#> 
#> $parameters
#> $parameters$a
#> [1] "myTable"
#> 
#> 

renderSql("SELECT * FROM @a {@b}?{WHERE x = 1};", a = "myTable", b = "true")

#> $originalSql
#> [1] "SELECT * FROM @a {@b}?{WHERE x = 1};"
#> 
#> $sql
#> [1] "SELECT * FROM myTable WHERE x = 1;"
#> 
#> $parameters
#> $parameters$a
#> [1] "myTable"
#> 
#> $parameters$b
#> [1] "true"
#> 
#> 

renderSql("SELECT * FROM @a {@b == ''}?{WHERE x = 1}:{ORDER BY x};", a = "myTable", b = "true")

#> $originalSql
#> [1] "SELECT * FROM @a {@b == ''}?{WHERE x = 1}:{ORDER BY x};"
#> 
#> $sql
#> [1] "SELECT * FROM myTable ORDER BY x;"
#> 
#> $parameters
#> $parameters$a
#> [1] "myTable"
#> 
#> $parameters$b
#> [1] "true"
#> 
#> 

renderSql("SELECT * FROM @a {@b != ''}?{WHERE @b = 1};", a = "myTable", b = "y")

#> $originalSql
#> [1] "SELECT * FROM @a {@b != ''}?{WHERE @b = 1};"
#> 
#> $sql
#> [1] "SELECT * FROM myTable WHERE y = 1;"
#> 
#> $parameters
#> $parameters$a
#> [1] "myTable"
#> 
#> $parameters$b
#> [1] "y"
#> 
#> 

renderSql("SELECT * FROM @a {1 IN (@c)}?{WHERE @b = 1};",
          a = "myTable",
          b = "y",
          c = c(1, 2, 3, 4))

#> $originalSql
#> [1] "SELECT * FROM @a {1 IN (@c)}?{WHERE @b = 1};"
#> 
#> $sql
#> [1] "SELECT * FROM myTable WHERE y = 1;"
#> 
#> $parameters
#> $parameters$a
#> [1] "myTable"
#> 
#> $parameters$b
#> [1] "y"
#> 
#> $parameters$c
#> [1] "1,2,3,4"
#> 
#> 

renderSql("{DEFAULT @b = \"someField\"}SELECT * FROM @a {@b != ''}?{WHERE @b = 1};",
          a = "myTable")

#> $originalSql
#> [1] "{DEFAULT @b = \"someField\"}SELECT * FROM @a {@b != ''}?{WHERE @b = 1};"
#> 
#> $sql
#> [1] "SELECT * FROM myTable WHERE someField = 1;"
#> 
#> $parameters
#> $parameters$a
#> [1] "myTable"
#> 
#> 

renderSql("SELECT * FROM @a {@a == 'myTable' & @b != 'x'}?{WHERE @b = 1};",
          a = "myTable",
          b = "y")

#> $originalSql
#> [1] "SELECT * FROM @a {@a == 'myTable' & @b != 'x'}?{WHERE @b = 1};"
#> 
#> $sql
#> [1] "SELECT * FROM myTable WHERE y = 1;"
#> 
#> $parameters
#> $parameters$a
#> [1] "myTable"
#> 
#> $parameters$b
#> [1] "y"
#> 
#>