snakeCaseToCamelCaseNames {SqlRender}	R Documentation

Convert the names of an object from snake case to camel case

Description

Convert the names of an object from snake case to camel case

Usage

snakeCaseToCamelCaseNames(object)

Arguments

object
The object of which the names should be converted

Value

The same object, but with converted names.

Examples

x <- data.frame(concept_id = 1, concept_name = "b")
snakeCaseToCamelCaseNames(x)

[Package SqlRender version 1.13.1 Index]