translateSql {SqlRender}	R Documentation

Deprecated: Translates SQL from one dialect to another

Description

This function has been deprecated. Use translate instead. This new function returns a character vector instead of a list.

Usage

translateSql(sql = "", targetDialect, oracleTempSchema = NULL)

Arguments

sql
The SQL to be translated

targetDialect
The target dialect. Currently "oracle", "postgresql", "pdw", "impala", "netezza", "bigquery", "snowflake", "synapse", "spark", and "redshift" are supported

oracleTempSchema
A schema that can be used to create temp tables in when using Oracle or Impala.

Value

A list containing the following elements:

originalSql

The original parameterized SQL code

sql

The translated SQL

[Package SqlRender version 1.13.1 Index]