renderSql {SqlRender}	R Documentation

Deprecated: Render SQL code based on parameterized SQL and parameter values

Description

This function has been deprecated. Use render instead. This new function returns a character vector instead of a list.

Usage

renderSql(sql = "", warnOnMissingParameters = TRUE, ...)

Arguments

sql
The parameterized SQL

warnOnMissingParameters
Should a warning be raised when parameters provided to this function do not appear in the parameterized SQL that is being rendered? By default, this is TRUE.

...
Parameter values

Value

A list containing the following elements:

parameterizedSql

The original parameterized SQL code

sql

The rendered sql

[Package SqlRender version 1.13.1 Index]